rust: Move Kconfig bool values into crate
Create a crate `zephyr-build` that contains the code that will run at
build time.  Move the bool kconfig handing from this.  This will make it
easier for an application crate that needs access to config entries to
do access them by:

- Adding a dependency to zephyr-build to `[build-dependencies]` in their
  Cargo.toml.
- Creating a build.rs, or adding a call to:
  zephyr_build::export_bool_kconfig() in it.
